requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-ninja==1.2.1
dnspython==2.6.1
email_validator==2.2.0
idna==3.7
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
